feat: Add userId field to Notes model and update createNote and deleteNote actions

+++ src/app/actions/deleteNote.ts
@@ -1,8 +1,13 @@
 "use server";
 
 import prisma from "@/server/db";
+import { auth } from "@clerk/nextjs/server";
 
 export async function deleteNote(id: number) {
+  const user = auth();
+  if (!user) {
+    return { success: false, error: "Not authenticated" };
+  }
   try {
     await prisma.notes.delete({
       where: {
